+++ packages/router-generator/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @tanstack/router-generator
 
+## 1.166.42
+
+### Patch Changes
+
+- Bump jiti dependency to 2.7.0. ([#7355](https://github.com/TanStack/router/pull/7355))
+
+## 1.166.41
+
+### Patch Changes
+
+- Updated dependencies [[`35e88f0`](https://github.com/TanStack/router/commit/35e88f04996d71019a1868b7b06ecb4ddbc9df9e)]:
+  - @tanstack/router-core@1.169.2
+
 ## 1.166.40
 
 ### Patch Changes