+++ packages/vue-start-client/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @tanstack/vue-start-client
 
+## 1.166.43
+
+### Patch Changes
+
+- Updated dependencies [[`35e88f0`](https://github.com/TanStack/router/commit/35e88f04996d71019a1868b7b06ecb4ddbc9df9e)]:
+  - @tanstack/router-core@1.169.2
+  - @tanstack/start-client-core@1.168.2
+  - @tanstack/vue-router@1.169.2
+
 ## 1.166.42
 
 ### Patch Changes